Added working directory on SMECY Internal Representation and #pragma to program and exchange between tools
This is my work in progress to define #pragma and API in SMECY: SME-C. :-)
This was presented in Delft on 2011/05/26.

So now all my SMECY information is pushed in the git and Janice should
have enough information to take over. :-)

There is also a SMECY directory in par4all-hpc-priv git with experiments
on applications we have to work on.

We should create a smecy git to collaborate between SMECY partners too.