Generate accelerator-side program skeleton with dispatch macros
Generate a ...-smecy_dispatch source file with some macros to implement
some dispatch code on the accelerator.

Move the initial main() to smecy_old_main() to avoid conflict in the
generated code.

Update generic user Makefile to produce accelerator-side code with
dispatching macros.

The MCAPI runtime implementation of the dispatch is to be done.